Refactor authMethods handling with, implement sync
* Implement `purgeUnmanagedConfig` option for auth methods in
bank-vaults cli-tool
* Refactor operator_client and split auth method logic
