+++ deploy/test-external-secrets-watch-deployment.yaml
@@ -90,8 +90,8 @@ spec:
     purgeUnmanagedConfig:
       enabled: true
       exclude:
-        # Auth is not implemented yet, but just for demonstration.
-        auth: true
+        # secrets is not implemented yet, but just for demonstration.
+        secrets: true
     policies:
       - name: allow_secrets
         rules: path "secret/*" {
